Update diff examples to Vite
